add endpoin get list tasks with ownerID

+++ src/tasks/tasks.controller.ts
@@ -1,13 +1,15 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, BadRequestException } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, BadRequestException, Query } from '@nestjs/common';
 import { TaskService } from '@/tasks/tasks.service';
 import { CreateTaskDto } from '@/tasks/dto/create-task.dto';
 import { UpdateTaskDto } from '@/tasks/dto/update-task.dto';
 import { User } from '@/decorators/user.decorator';
 import type { IUser } from '@/users/interfaces/IUser';
+import type { IFilterDate } from '@/tasks/interfaces/IFilterDate';
+import { ApiQuery } from '@nestjs/swagger';
 
 @Controller('tasks')
 export class TaskController {
-  constructor(private readonly taskService: TaskService) {}
+  constructor(private readonly taskService: TaskService) { }
 
   @Post()
   async create(@Body() createTaskDto: CreateTaskDto, @User() user: IUser) {
@@ -19,13 +21,49 @@ export class TaskController {
   }
 
   @Get()
-  findAll() {
-    return this.taskService.findAll();
+  @ApiQuery({ name: 'filterDate', required: false, enum: ['all_time', 'today', 'this_week', 'this_month', 'this_year'], default: 'all_time' })
+  async findAll(@User() user: IUser, @Query('filterDate') filterDate?: IFilterDate) {
+    filterDate ? filterDate.toString() : 'all_time';
+    const now = new Date();
+    let startDate: Date | null = null;
+    switch (filterDate) {
+      case 'today': {
+        startDate = new Date(now.getFullYear(), now.getMonth(), now.getDate()); // 11/08/2024 00:00:00
+        break;
+      };
+      case 'this_week': {
+        const mondayDate = now.getDate() - (now.getDay() - 1) - (now.getDay() === 0 ? 7 : 0);
+        startDate = new Date(now.getFullYear(), now.getMonth(), mondayDate); // Set to the most recent Monday
+        break;
+      };
+      case 'this_month': {
+        startDate = new Date(now.getFullYear(), now.getMonth(), 1);
+        break;
+      };
+      case 'this_year': {
+        startDate = new Date(now.getFullYear(), 0, 1);
+        break;
+      };
+      case 'all_time':
+      default: {
+        startDate = null; // No date filtering
+        break;
+      };
+    };
+    const results = await this.taskService.findAllTasksByOwnerId(user._id, startDate);
+    if( results.tasks.length === 0 ){
+      return { message: 'No tasks found', data: results };
+    }
+    return { message: 'Tasks retrieved successfully', data: results };
   }
 
   @Patch(':id')
   async update(@Param('id') id: string, @Body() updateTaskDto: UpdateTaskDto, @User() user: IUser) {
     try {
+      const isOwner = await this.taskService.TaskIsOwner(id, user._id);
+      if (!isOwner) {
+        throw new BadRequestException('You do not have permission to update this task');
+      }
       return { message: 'Task updated successfully', data: await this.taskService.update(id, updateTaskDto, user._id) };
     } catch (error) {
       throw new BadRequestException('Failed to update task: ' + error.message);
@@ -35,6 +73,10 @@ export class TaskController {
   @Delete(':id')
   async remove(@Param('id') id: string, @User() user: IUser) {
     try {
+      const isOwner = await this.taskService.TaskIsOwner(id, user._id);
+      if (!isOwner) {
+        throw new BadRequestException('You do not have permission to remove this task');
+      }
       return { message: 'Task removed successfully', data: await this.taskService.remove(id, user._id) };
     } catch (error) {
       throw new BadRequestException('Failed to remove task: ' + error.message);
